modify vuex
