+++ src/store/getters.js
@@ -1,0 +1,5 @@
+export default {
+  getCurrentUser: (state) => state.currentUser,
+  getViewUser: (state) => state.viewUser,
+  getPopular: (state) => state.popular,
+}
