+++ src/views/User/UserAllFollow.vue
@@ -9,17 +9,19 @@
       <router-link
         :to="{ name: 'UserFollowers', params: { user_id: userId } }"
         class="tab-router__link"
+        :class="{ 'router-link-active': tab === 'UserFollowers' }"
       >
         <span class="tab-router__text">跟隨者</span>
       </router-link>
       <router-link
         :to="{ name: 'UserFollowings', params: { user_id: userId } }"
         class="tab-router__link"
+        :class="{ 'router-link-active': tab === 'UserFollowings' }"
       >
         <span class="tab-router__text">正在跟隨</span>
       </router-link>
     </section>
-    <router-view />
+    <router-view :isLoadingPage="isLoading" />
   </div>
 </template>
 
@@ -48,20 +50,23 @@ export default {
     currentViewUser() {
       return this.$store.getters.getViewUser
     },
+    tab() {
+      return this.$route.name
+    },
   },
   created() {
     const { user_id } = this.$route.params
     this.userId = user_id
     this.fetchUser(user_id)
+    this.$store.dispatch('isViewCurrentUser', user_id)
     this.$store.dispatch('fetchViewUserFollowings')
     this.$store.dispatch('fetchViewUserFollowers')
   },
   beforeRouteUpdate(to, from, next) {
-    const { name } = this.$route
-    this.show = name
     const { user_id } = this.$route.params
     this.userId = user_id
     this.fetchUser(user_id)
+    this.$store.dispatch('isViewCurrentUser', user_id)
     this.$store.dispatch('fetchViewUserFollowings')
     this.$store.dispatch('fetchViewUserFollowers')
     next()
@@ -114,7 +119,8 @@ export default {
     &:hover {
       color: var(--theme);
     }
-    &.router-link-active {
+    &.router-link-active,
+    .router-link-exact-active {
       color: var(--theme);
       border-bottom: 2px solid var(--theme);
     }
